requirements.txt
websockets
soundfile
sounddevice
numpy
python-dotenv